refactoring to use BigDecimal instead of double or int

+++ src/main/java/cucumbercourse/model/BillAmount.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class BillAmount {
-    private double billAmount;
-    private double taxRate;
-    private double billCalculated;
+    private BigDecimal billAmount;
+    private BigDecimal taxRate;
+    private BigDecimal billCalculated;
 }
